requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
asyncpg==0.31.0
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.6.4
frozenlist==1.8.0
greenlet==3.3.1
idna==3.11
librt==0.7.8
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
pathspec==1.0.4
propcache==0.4.1
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
yarl==1.22.0
